test-requirements.txt
charms.reactive
flake8>=2.2.4
mock>=1.2
stestr>=2.2.0
git+https://github.com/openstack/charms.openstack.git#egg=charms.openstack